DwarfTherapist is a new QApplication subclass that handles global app data from now on.
Overhaul to the options menu and how settings are propagated. There is a new global macro "DT" for getting a reference to the global DwarfTherapist app. From there you can get the user settings object.

Starting to come together a bit more. The tabs are working by showing labors with the active flag set to true or 1.

git-svn-id: http://dwarftherapist.googlecode.com/svn/trunk@99 f34a0644-6cda-11de-9313-dd366511c2ff